패킷 분배기 코드는 클라/서버에서 공유하기 bot을 console로 포팅

+++ Hatsushimo/NetChan/PacketEnumerator.cs
@@ -8,7 +8,7 @@ namespace Hatsushimo.NetChan
         readonly PacketCodec codec = new PacketCodec();
 
         short currType;
-        public int CurrentType { get { return currType; } }
+        public short CurrentType { get { return currType; } }
 
         public PacketEnumerator(BinaryReader reader)
         {
